Issue #48: Add support for MongoDB
Split the core-context.xml and use different files for each message
store.
Create a MongoMessageStore class.
Add some tests for the new MongoMessageStore.
Change the message id to an object instead of a long.
Fix some issues with the id change.

+++ mokai-core/src/test/java/org/mokai/persist/jdbc/test/JdbcMessageStoreTest.java
@@ -45,7 +45,7 @@ public class JdbcMessageStoreTest {
 		Message message = new Message();
 		messageStore.saveOrUpdate(message);
 		
-		Assert.assertEquals(10, message.getId());
+		Assert.assertEquals(10L, ((Long) message.getId()).longValue());
 		
 		verify(handler).insertMessage(any(Connection.class), any(Message.class));
 		verify(handler, never()).updateMessage(any(Connection.class), any(Message.class));
